Moved UIImage extension to it's own file.
